+++ blink_clip_downloader/blink_downloader/__init__.py
@@ -1,3 +1,3 @@
 """Blink Clip Downloader - Home Assistant add-on."""
 
-__version__ = "5.4.4"
+__version__ = "5.4.5"